server-old-v1/requirements.txt
pytz
discord.py[voice]
mysql-connector-python